move nearest_...h to Kernel_23
